requirements.txt
asgiref==3.8.1
Django==5.1.7
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
inflection==0.5.1
Markdown==3.7
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
uuid==1.30
